This is the new approach for a comments feature. Implements #51.
This is based on a re-design from @halprin's code review in #52.
