+++ fc/cli/main.py
@@ -4,7 +4,7 @@ from . import backlog
 from . import el
 from . import version
 from . import task
-
+from . import comment
 
 @click.group()
 def cli():
@@ -16,3 +16,4 @@ cli.add_command(backlog.backlog)
 cli.add_command(el.el)
 cli.add_command(version.version)
 cli.add_command(task.task)
+cli.add_command(comment.comment)
